+++ gulpfile.ts
@@ -2,6 +2,8 @@ import fs from "fs";
 import gulp from "gulp";
 import gulpTypescript from "gulp-typescript";
 import gulpUglify from "gulp-uglify";
+import webpackStream from "webpack-stream";
+import webpackConfig from "./webpack.config.js";
 
 const TSCONFIG_JSON: string = "tsconfig.json";
 let tsConfig: any;
@@ -16,8 +18,8 @@ try {
 const TS_PROJECT: gulpTypescript.Project = gulpTypescript.createProject(TSCONFIG_JSON);
 const DIR_OUT: string = tsConfig.compilerOptions.outDir;
 
-export default compile;
+export default build;
 
-export async function compile(): Promise<NodeJS.ReadWriteStream> {
-	return TS_PROJECT.src().pipe(TS_PROJECT()).js.pipe(gulpUglify()).pipe(gulp.dest(DIR_OUT));
+export async function build(): Promise<NodeJS.ReadWriteStream> {
+	return TS_PROJECT.src().pipe(TS_PROJECT()).js.pipe(gulpUglify()).pipe(webpackStream(webpackConfig)).pipe(gulp.dest(DIR_OUT));
 }