+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 # CHANGELOG
 ## [Unreleased]
+- webpack.config.js: cjs -> ejs